Revert "refactor: extract diff file parsing"
This reverts commit a7f5e1ba6bde2a9ea5d1dabb4dba99e2c3db78ac.
