+++ apps/api/src/utils/git.ts
@@ -1,4 +1,4 @@
-import { BranchRef, DiffMode } from "@diffview/shared";
+import { BranchRef, ChangedFile, DiffMode } from "@diffview/shared";
 import { spawn } from "bun";
 
 async function execGit(args: string[], allowExitCodes: number[] = [0]): Promise<string> {
@@ -101,6 +101,42 @@ export async function getDiffPatches(
   return execGit(args);
 }
 
+export function parseDiffFiles(rawList: string[]): ChangedFile[] {
+  return rawList.map((line) => {
+    const parts = line.split("\t");
+    const statusCode = parts[0];
+    const status =
+      statusCode === "A" || statusCode === "??"
+        ? "added"
+        : statusCode === "M"
+          ? "modified"
+          : statusCode === "D"
+            ? "deleted"
+            : statusCode.startsWith("R")
+              ? "renamed"
+              : "modified";
+
+    if (parts.length === 3) {
+      return {
+        path: parts[2],
+        oldPath: parts[1],
+        status,
+      };
+    }
+
+    if (parts.length === 2) {
+      return {
+        path: parts[1],
+        status,
+      };
+    }
+
+    return {
+      path: parts[0],
+      status,
+    };
+  });
+}
 
 export function splitPatchByFile(patch: string): Map<string, string> {
   const files = new Map<string, string>();
